Implement feedback

+++ src/test/java/de/tum/in/www1/artemis/hestia/ExerciseHintServiceTest.java
@@ -176,7 +176,7 @@ public class ExerciseHintServiceTest extends AbstractSpringIntegrationBambooBitb
 
     @Test
     public void testGetAvailableExerciseHintsWithZeroThreshold1() {
-        hints.get(0).setDisplayThreshold(0);
+        hints.get(0).setDisplayThreshold((short) 0);
         exerciseHintRepository.save(hints.get(0));
         addResultWithFailedTestCases(exercise.getTestCases());
         var availableExerciseHints = exerciseHintService.getAvailableExerciseHints(exercise, student);
@@ -185,7 +185,7 @@ public class ExerciseHintServiceTest extends AbstractSpringIntegrationBambooBitb
 
     @Test
     public void testGetAvailableExerciseHintsWithZeroThreshold2() {
-        hints.get(0).setDisplayThreshold(0);
+        hints.get(0).setDisplayThreshold((short) 0);
         exerciseHintRepository.save(hints.get(0));
         addResultWithSuccessfulTestCases(exercise.getTestCases());
         var availableExerciseHints = exerciseHintService.getAvailableExerciseHints(exercise, student);
@@ -194,7 +194,7 @@ public class ExerciseHintServiceTest extends AbstractSpringIntegrationBambooBitb
 
     @Test
     public void testGetAvailableExerciseHintsWithZeroThreshold3() {
-        hints.get(0).setDisplayThreshold(0);
+        hints.get(0).setDisplayThreshold((short) 0);
         exerciseHintRepository.save(hints.get(0));
         addResultWithFailedTestCases(exercise.getTestCases());
         addResultWithSuccessfulTestCases(sortedTasks.get(0).getTestCases());
